rag-chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
google-generativeai